Renamed package common -> error
