+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/dccreissuance/core/error/DccReissuanceException.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.dccreissuance.core.common
+package de.rki.coronawarnapp.dccreissuance.core.error
 
 import android.content.Context
 import de.rki.coronawarnapp.R
