+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/dccreissuance/core/server/validation/DccReissuanceServerCertificateValidator.kt
@@ -1,7 +1,7 @@
 package de.rki.coronawarnapp.dccreissuance.core.server.validation
 
 import de.rki.coronawarnapp.appconfig.AppConfigProvider
-import de.rki.coronawarnapp.dccreissuance.core.common.DccReissuanceException
+import de.rki.coronawarnapp.dccreissuance.core.error.DccReissuanceException
 import de.rki.coronawarnapp.appconfig.CovidCertificateConfig
 import de.rki.coronawarnapp.tag
 import okio.ByteString.Companion.toByteString
